fix(vscode-server): remove refactored code

+++ output/rstudio/start-custom.sh
@@ -86,7 +86,6 @@ printenv | grep KUBERNETES >> /opt/conda/lib/R/etc/Renviron
 
 # persist vscode server remote settings (Machine dir)
 VS_CODE_SETTINGS=${XDG_DATA_HOME}/code-server/Machine/settings.json
-# if [! -f "$VS_CODE_SETTINGS" ]; then
-#     cp vscode-overrides.json "$VS_CODE_SETTINGS"
-# fi
-cp vscode-overrides.json "$VS_CODE_SETTINGS"
+if [! -f "$VS_CODE_SETTINGS" ]; then
+    cp vscode-overrides.json "$VS_CODE_SETTINGS"
+fi
